[Backend] Split ModelBuilder and ResourceBuilder
